consumer/requirements.txt
elastic-transport==8.1.2
elasticsearch==8.2.0 
iocextract==1.13.1
kafka-python==2.0.2
nltk==3.7
regex==2022.3.15
textblob==0.17.1
transformers==4.17.0